Added _USERNAME namespacing to avoid dev conflicts
I realized as I was working on a shared workstation that my docker
environment's user was `mrsj` (Matthew) and not `voigte` (me). This was
a bug that I have (hopefully) fixed with this commit.

It requires that users copy docker-compose.yml.dist into
docker-compose.yml and do a few changes (replacing `USERNAME` with their
username). the namespacing helps avoid conflicts.

Also switched base app image to osuosl/centos-ruby